+++ apps/vaults-v3/components/details/VaultDetailsHeader.tsx
@@ -7,6 +7,7 @@ import {toBigInt, toNormalizedBN} from '@yearn-finance/web-lib/utils/format.bigN
 import {formatUSD} from '@yearn-finance/web-lib/utils/format.number';
 import {formatCounterValue} from '@yearn-finance/web-lib/utils/format.value';
 import {copyToClipboard} from '@yearn-finance/web-lib/utils/helpers';
+import {isZero} from '@yearn-finance/web-lib/utils/isZero';
 import {getNetwork} from '@yearn-finance/web-lib/utils/wagmi/utils';
 import {RenderAmount} from '@common/components/RenderAmount';
 import {useBalance} from '@common/hooks/useBalance';
@@ -54,6 +55,9 @@ function VaultAPR({apr}: {apr: TYDaemonVault['apr']}): ReactElement {
 			</VaultHeaderLineItem>
 		);
 	}
+
+	const monthlyAPR = apr.netAPR + apr.extra.stakingRewardsAPR;
+	const weeklyAPR = apr.points.weekAgo;
 	return (
 		<VaultHeaderLineItem
 			label={'Historical APR'}
@@ -89,7 +93,7 @@ function VaultAPR({apr}: {apr: TYDaemonVault['apr']}): ReactElement {
 				shouldRender={!apr?.type.includes('new')}
 				fallback={'New'}>
 				<RenderAmount
-					value={apr?.netAPR + apr.extra.stakingRewardsAPR}
+					value={isZero(monthlyAPR) ? weeklyAPR : monthlyAPR}
 					symbol={'percent'}
 					decimals={6}
 				/>