+++ apps/vaults/components/list/VaultsListRow.tsx
@@ -251,6 +251,8 @@ export function VaultForwardAPR({currentVault}: {currentVault: TYDaemonVault}): 
 function VaultHistoricalAPR({currentVault}: {currentVault: TYDaemonVault}): ReactElement {
 	const hasZeroAPR =
 		isZero(currentVault.apr?.netAPR) || Number(Number(currentVault.apr?.netAPR || 0).toFixed(2)) === 0;
+	const monthlyAPR = currentVault.apr.points.monthAgo;
+	const weeklyAPR = currentVault.apr.points.weekAgo;
 
 	if (currentVault.apr?.extra.stakingRewardsAPR > 0) {
 		return (
@@ -266,7 +268,7 @@ function VaultHistoricalAPR({currentVault}: {currentVault: TYDaemonVault}): Reac
 								}>
 								<RenderAmount
 									shouldHideTooltip={hasZeroAPR}
-									value={currentVault.apr?.netAPR}
+									value={isZero(monthlyAPR) ? weeklyAPR : monthlyAPR}
 									symbol={'percent'}
 									decimals={6}
 								/>
@@ -286,7 +288,7 @@ function VaultHistoricalAPR({currentVault}: {currentVault: TYDaemonVault}): Reac
 									<p>{'• Base APR '}</p>
 									<RenderAmount
 										shouldHideTooltip
-										value={currentVault.apr.netAPR}
+										value={isZero(monthlyAPR) ? weeklyAPR : monthlyAPR}
 										symbol={'percent'}
 										decimals={6}
 									/>
@@ -314,7 +316,7 @@ function VaultHistoricalAPR({currentVault}: {currentVault: TYDaemonVault}): Reac
 					shouldRender={!currentVault.apr?.type.includes('new')}
 					fallback={'NEW'}>
 					<RenderAmount
-						value={currentVault.apr?.netAPR}
+						value={isZero(monthlyAPR) ? weeklyAPR : monthlyAPR}
 						shouldHideTooltip={hasZeroAPR}
 						symbol={'percent'}
 						decimals={6}