+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 -   Feat: Update the vault schema to take into account the new V3 composite elements.
 -   Feat: For the V3 vaults, display the spot expected APR next to the Net APR.
+-   Feat: Display the weekly APR when the Monthly APR is zero for historical.
 
 # 0.3.4 (22/12/2023)